[DDW-1012] Stake pool list performance
