+++ package.json
@@ -103,6 +103,7 @@
     "@types/qrcode.react": "^1.0.2",
     "@types/react": "^17.0.38",
     "@types/react-svg-inline": "^2.1.3",
+    "@types/react-table": "^7.7.9",
     "@typescript-eslint/eslint-plugin": "^5.10.1",
     "@typescript-eslint/parser": "^5.10.1",
     "asar": "2.1.0",
@@ -260,6 +261,7 @@
     "react-router": "5.2.0",
     "react-router-dom": "5.2.0",
     "react-svg-inline": "2.1.1",
+    "react-table": "7.7.0",
     "react-virtualized": "9.22.3",
     "recharts": "1.8.5",
     "rimraf": "3.0.2",
