+++ source/renderer/app/components/staking/stake-pools/StakePoolsTable.scss
@@ -92,12 +92,12 @@
     }
   }
 
-  table {
+  .table {
     border-style: hidden;
     user-select: text;
     width: 100%;
 
-    thead {
+    .thead {
       background: var(--theme-bordered-box-background-color);
       border-radius: 4px 4px 0 0;
       box-shadow: 0 10px 10px -10px rgba(0, 0, 0, 0.2);
@@ -108,7 +108,7 @@
       top: 0;
       z-index: $sticky-header-z-index;
 
-      tr {
+      .tr {
         border: 0;
         display: flex;
         justify-content: space-between;
@@ -131,8 +131,8 @@
       }
     }
 
-    tbody {
-      tr {
+    .tbody {
+      .tr {
         display: flex;
         justify-content: space-between;
         width: 100%;
@@ -145,11 +145,11 @@
       }
     }
 
-    tr {
+    .tr {
       border-bottom: 1px solid var(--theme-staking-table-border-color);
     }
 
-    th {
+    .th {
       color: var(--theme-staking-font-color-regular);
       cursor: pointer;
       font-family: var(--font-semibold);
@@ -207,7 +207,7 @@
       }
     }
 
-    td {
+    .td {
       color: var(--theme-staking-font-color-regular);
       font-family: var(--font-regular);
       font-size: 12px;
@@ -266,6 +266,7 @@
         display: flex;
         flex-direction: row;
         height: 100%;
+        justify-content: center;
       }
 
       %saturationContainer {
@@ -346,8 +347,8 @@
       }
     }
 
-    th,
-    td {
+    .th,
+    .td {
       &:nth-child(1),
       &:nth-child(3),
       &:nth-child(4),
@@ -363,39 +364,39 @@
 
       // Rank
       &:nth-child(1) {
-        width: 7%;
+        width: 7% !important;
       }
       // Ticker
       &:nth-child(2) {
-        width: 9%;
+        width: 9% !important;
       }
       // Saturation
       &:nth-child(3) {
-        width: 11%;
+        width: 11% !important;
       }
       // Cost
       &:nth-child(4) {
-        width: 10%;
+        width: 10% !important;
       }
       // Margin
       &:nth-child(5) {
-        width: 8%;
+        width: 8% !important;
       }
       // Produced Blocks
       &:nth-child(6) {
-        width: 15%;
+        width: 15% !important;
       }
       // Potential Rewards
       &:nth-child(7) {
-        width: 16%;
+        width: 16% !important;
       }
       // Pledge
       &:nth-child(8) {
-        width: 12%;
+        width: 12% !important;
       }
       // Retiring in
       &:nth-child(9) {
-        width: 12%;
+        width: 12% !important;
       }
     }
   }
