+++ storybook/webpack.config.js
@@ -42,10 +42,10 @@ module.exports = async ({ config }) => {
             'moment',
             'pbkdf2',
             'qrcode.react',
-            'react',
+            // 'react',
             'react-copy-to-clipboard',
             'react-datetime',
-            'react-dom',
+            // 'react-dom',
             'react-router',
             'react-svg-inline',
             'recharts',
